+++ include/compute/compute_partial_derivatives.hpp
@@ -1,7 +1,6 @@
 #ifndef _COMPUTE_PARTIAL_DERIVATIVES_HPP
 #define _COMPUTE_PARTIAL_DERIVATIVES_HPP
 
-#include "compute/compute_mesh.hpp"
 #include "enumerations/specfem_enums.hpp"
 #include "kokkos_abstractions.h"
 #include "macros.hpp"
