+++ include/compute/properties/properties.hpp
@@ -1,11 +1,12 @@
-#pragma once
+#ifndef _COMPUTE_PROPERTIES_HPP
+#define _COMPUTE_PROPERTIES_HPP
 
 #include "enumerations/specfem_enums.hpp"
 #include "impl/material_properties.hpp"
 #include "impl/properties_container.hpp"
 #include "kokkos_abstractions.h"
 #include "macros.hpp"
-#include "medium/material.hpp"
+#include "material/material.hpp"
 #include "point/coordinates.hpp"
 #include "specfem_setup.hpp"
 #include <Kokkos_Core.hpp>
@@ -285,3 +286,5 @@ void store_on_host(const IndexType &lcoord,
 }
 } // namespace compute
 } // namespace specfem
+
+#endif
