+++ include/mesh/materials/materials.tpp
@@ -1,7 +1,8 @@
-#pragma once
+#ifndef _MESH_MATERIALS_TPP
+#define _MESH_MATERIALS_TPP
 
 #include "enumerations/specfem_enums.hpp"
-#include "medium/material.hpp"
+#include "material/material.hpp"
 #include "materials.hpp"
 #include <variant>
 #include <vector>
@@ -10,7 +11,9 @@ template <specfem::element::medium_tag type,
           specfem::element::property_tag property>
 specfem::mesh::materials::material<type, property>::material(
     const int n_materials,
-    const std::vector<specfem::medium::material<type, property> >
+    const std::vector<specfem::material::material<type, property> >
         &l_materials)
     : n_materials(n_materials),
       material_properties(l_materials) {}
+
+#endif /* end of include guard: _MESH_MATERIALS_TPP */
