+++ tests/unit-tests/CMakeLists.txt
@@ -109,6 +109,7 @@ target_link_libraries(
   kokkos_environment
   yaml-cpp
   IO
+  # material_class
   -lpthread -lm
 )
 
@@ -127,6 +128,7 @@ target_link_libraries(
   yaml-cpp
   compare_arrays
   Boost::filesystem
+  # material_class
   -lpthread -lm
 )
 
@@ -145,6 +147,7 @@ target_link_libraries(
   yaml-cpp
   compare_arrays
   Boost::filesystem
+  # material_class
   -lpthread -lm
 )
 
@@ -162,6 +165,7 @@ target_link_libraries(
 #   kokkos_environment
 #   yaml-cpp
 #   compare_arrays
+#   # material_class
 #   -lpthread -lm
 # )
 
@@ -181,6 +185,7 @@ target_link_libraries(
   compare_arrays
   IO
   Boost::filesystem
+  # material_class
   -lpthread -lm
 )
 
@@ -267,6 +272,7 @@ target_link_libraries(
 
 # # target_link_libraries(
 # #   source_location_tests
+# #   material_class
 # #   mpi_environment
 # #   kokkos_environment
 # #   yaml-cpp
@@ -286,6 +292,7 @@ target_link_libraries(
   rmass_inverse_tests
   quadrature
   mesh
+  # material_class
   yaml-cpp
   kokkos_environment
   mpi_environment
@@ -310,6 +317,7 @@ target_link_libraries(
   displacement_newmark_tests
   quadrature
   mesh
+  # material_class
   yaml-cpp
   kokkos_environment
   mpi_environment
